update puzzlink encoder
